move ErrorBoundary and PrettyCodeFormatter to /components (new dir) and rearrange args to parseIntWithFallback function
