+++ src/common/PrettyCodeFormatter.js
