+++ src/utils/nutils.js
@@ -1,5 +1,5 @@
 
-export const parseIntWithFallback = (num, radix = 10, fallback = 0) => {
+export const parseIntWithFallback = (num, fallback = 0, radix = 10,) => {
   // a simple function that allows you to specify a fallback in case
   // the parseInt() call returns a falsy value (NaN)
   return parseInt(num, radix) || fallback;
